FIX: fix two bugs handling empty strings
-ascii_string correctly handles `''` and does not need to be
 short-cirucited

- if the length of the input is 0, then the length is re-set to the
maximum length of the array.  Later, there is a check that if ftype ==
dbr.CHAR then add 1 to count.  This results in a error from trying to
put a too-long array to the PV.

There are still bugs in handling the edge cases of very long
strings (the null char might not get written).