CI: tweak how docker is started
This will make it possible to start more than one docker instance
and have them live side-by-side.

This binds each servers 5064 tcp port to a random port in the range
7000-8000 which is sorted out by libca as part of the search process.

The udp multi-cast passes through on 5064 (as it did before).

We do not need to expose the heart-beat outside of the docker manager.